@@ -9,6 +9,10 @@ logger = Logger()
 
 class ElasticLoadBalancer:
 
+    @property
+    def instances(self) -> list:
+        return self.api_response['Instances']
+
     @staticmethod
     def find_elastic_load_balancers(names:list) -> list:
         return list(map(lambda load_balancer: ElasticLoadBalancer(load_balancer),
@@ -22,24 +26,22 @@ class ElasticLoadBalancer:
             if not load_balancers:
                 logger.error("Mode not enabled. ETL load balancer doesn't exist.")
                 raise Exception('No ETL load balancer found!')
+            # logger.info(f"Using load balancer name: {load_balancers}")
         except ClientError as e:
             if e.response['Error']['Code'] == 'LoadBalancerNotFound':
                 return None
             else:
                 raise e
         return load_balancers[0]
-    
+
     def __init__(self, api_response):
         self.api_response = api_response
+
 
     @cached_property
     def load_balancer_name(self):
         return self.api_response['LoadBalancerName']
-    
-    @property
-    def instances(self) -> list:
-        return self.api_response['Instances']
-    
+
     def reload(self):
         self.api_response = list(filter(lambda load_balancer: load_balancer['LoadBalancerName'] == self.load_balancer_name, 
                                         ElasticLoadBalancerGateway.describe_load_balancers([self.load_balancer_name])))[0]
@@ -58,3 +60,6 @@ class ElasticLoadBalancer:
     def deregister_instances(self, instances:list):
         ElasticLoadBalancerGateway.deregister_instances_from_load_balancer(self.load_balancer_name, instances)
         self.reload()
+
+    def __repr__(self):
+        return f"<ElasticLoadBalancer(name={self.load_balancer_name}, instances={len(self.instances)})>"
